remove annotation
