+++ python/tests/engine/test_python.py
@@ -3767,7 +3767,7 @@ class TestPython:
         mocker.patch("hsfs.util.get_job_url")
         mocker.patch(
             "hsfs.core.kafka_engine.kafka_get_offsets",
-            side_effect=["tests_offsets1", "tests_offsets1", "tests_offsets2"],
+            side_effect=["tests_offsets1", "tests_offsets2"],
         )
         mock_fg_api = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi")
         mocker.patch(
